Fix modify interface logic

+++ worker/modify_interface_module/restconf_client_write.py
@@ -35,15 +35,18 @@ def configure_interface(
     }
 
     try:
-        requests.delete(url, auth=(username, password), verify=False)
-        response = requests.put(
-            url,
-            headers=headers,
-            auth=(username, password),
-            data=json.dumps(payload),
-            verify=False,
-        )
-        response.raise_for_status()
+        if (iface_ip == "" or iface_mask == "") or (iface_ip == router_ip):
+            return True, "Configuration successful"
+        else:
+            requests.delete(url, auth=(username, password), verify=False)
+            response = requests.put(
+                url,
+                headers=headers,
+                auth=(username, password),
+                data=json.dumps(payload),
+                verify=False,
+            )
+            response.raise_for_status()
         return True, "Configuration successful"
 
     except requests.exceptions.RequestException as e:
